Fix bug with gas price calculation, fixed tx popup for proposing milestone, moved confirmtx to middleware
