+++ src/components/UnlockWalletForm.jsx
@@ -47,7 +47,7 @@ class UnlockWalletForm extends Component {
           <div className="form-group">
             <Input
               autoFocus
-              autocomplete="current-password"
+              autoComplete="current-password"
               value={password}
               name="password"
               id="password-input"
