+++ src/components/WithdrawButton.jsx
@@ -3,11 +3,11 @@ import PropTypes from 'prop-types';
 import { SkyLightStateless } from 'react-skylight';
 import { Form, Input } from 'formsy-react-components';
 
-import { takeActionAfterWalletUnlock } from '../lib/middleware';
+import { takeActionAfterWalletUnlock, confirmBlockchainTransaction } from '../lib/middleware';
 import User from '../models/User';
 import GivethWallet from '../lib/blockchain/GivethWallet';
 import WalletService from '../services/Wallet';
-import { getGasPrice, confirmBlockchainTransaction } from '../lib/helpers';
+import { getGasPrice } from '../lib/helpers';
 
 class WithdrawButton extends Component {
   constructor() {
