+++ src/components/views/EditMilestone.jsx
@@ -11,9 +11,9 @@ import FormsyImageUploader from './../FormsyImageUploader';
 import GoBackButton from '../GoBackButton';
 import {
   isOwner, displayTransactionError, getRandomWhitelistAddress, getTruncatedText,
-  confirmBlockchainTransaction, getGasPrice,
+  getGasPrice,
 } from '../../lib/helpers';
-import { isAuthenticated, checkWalletBalance, isInWhitelist } from '../../lib/middleware';
+import { isAuthenticated, checkWalletBalance, isInWhitelist, confirmBlockchainTransaction } from '../../lib/middleware';
 import getNetwork from '../../lib/blockchain/getNetwork';
 import getWeb3 from '../../lib/blockchain/getWeb3';
 import LoaderButton from '../../components/LoaderButton';
@@ -203,17 +203,33 @@ class EditMilestone extends Component {
       }
     };
 
-    // Save the Milestone
-    confirmBlockchainTransaction(
-      () => {
-        if (this.state.uploadNewImage) {
-          feathersClient.service('/uploads').create({ uri: this.state.image }).then(file => updateMilestone(file.url));
-        } else {
-          updateMilestone();
-        }
-      },
-      () => this.setState({ isSaving: false }),
-    );
+    const saveMilestone = () => {
+      if (this.state.uploadNewImage) {
+        feathersClient.service('/uploads').create({ uri: this.state.image }).then(file => updateMilestone(file.url));
+      } else {
+        updateMilestone();
+      }
+    }    
+
+    if(this.props.isProposed) {
+      React.swal({
+        title: 'Propose milestone?',
+        text:
+          'The milestone will be proposed to the campaign owner and he or she might approve or reject your milestone.',
+        icon: 'warning',
+        dangerMode: true,
+        buttons: ['Cancel', 'Yes, propose'],
+      }).then(isConfirmed => {
+        if (isConfirmed) saveMilestone()
+      }); 
+    } else {
+      // Save the Milestone
+      confirmBlockchainTransaction(
+        saveMilestone(),
+        () => this.setState({ isSaving: false }),
+        this.props.isProposed
+      );
+    }
   }
 
   toggleFormValid(state) {
